// libs/importer/src/importers/keeper/access/generated/SyncDown.ts
export interface UserFolder {
  /**
   * @generated from protobuf field: bytes folderUid = 1
   */
  folderUid: Uint8Array;
  /**
   * @generated from protobuf field: bytes parentUid = 2
   */
  parentUid: Uint8Array;
  /**
   * @generated from protobuf field: bytes userFolderKey = 3
   */
  userFolderKey: Uint8Array;
  /**
   * @generated from protobuf field: Records.RecordKeyType keyType = 4
   */
  keyType: RecordKeyType;
  /**
   * @generated from protobuf field: int64 revision = 5
   */
  revision: bigint;
  /**
   * @generated from protobuf field: bytes data = 6
   */
  data: Uint8Array;
}
/**
 * @generated from protobuf message Vault.SharedFolder
 */
export interface SharedFolder {
  /**
   * @generated from protobuf field: bytes sharedFolderUid = 1
   */
  sharedFolderUid: Uint8Array;
  /**
   * @generated from protobuf field: int64 revision = 2
   */
  revision: bigint;
  /**
   * @generated from protobuf field: bytes sharedFolderKey = 3
   */
  sharedFolderKey: Uint8Array;
  /**
   * @generated from protobuf field: Records.RecordKeyType keyType = 4
   */
  keyType: RecordKeyType;
  /**
   * @generated from protobuf field: bytes data = 5
   */
  data: Uint8Array;
  /**
   * @generated from protobuf field: bool defaultManageRecords = 6
   */
  defaultManageRecords: boolean;
  /**
   * @generated from protobuf field: bool defaultManageUsers = 7
   */
  defaultManageUsers: boolean;
  /**
   * @generated from protobuf field: bool defaultCanEdit = 8
   */
  defaultCanEdit: boolean;
  /**
   * @generated from protobuf field: bool defaultCanReshare = 9
   */
  defaultCanReshare: boolean;
  /**
   * @generated from protobuf field: Vault.CacheStatus cacheStatus = 10
   */
  cacheStatus: CacheStatus;
  /**
   * @generated from protobuf field: string owner = 11
   */
  owner: string;
  /**
   * @generated from protobuf field: bytes ownerAccountUid = 12
   */
  ownerAccountUid: Uint8Array;
  /**
   * @generated from protobuf field: bytes name = 13
   */
  name: Uint8Array;
}
// ...
export interface Record {
  /**
   * @generated from protobuf field: bytes recordUid = 1
   */
  recordUid: Uint8Array;
  /**
   * @generated from protobuf field: int64 revision = 2
   */
  revision: bigint;
  /**
   * @generated from protobuf field: int32 version = 3
   */
  version: number;
  /**
   * @generated from protobuf field: bool shared = 4
   */
  shared: boolean;
  /**
   * @generated from protobuf field: int64 clientModifiedTime = 5
   */
  clientModifiedTime: bigint;
  /**
   * @generated from protobuf field: bytes data = 6
   */
  data: Uint8Array;
  /**
   * @generated from protobuf field: bytes extra = 7
   */
  extra: Uint8Array;
  /**
   * @generated from protobuf field: string udata = 8
   */
  udata: string;
  /**
   * @generated from protobuf field: int64 fileSize = 9
   */
  fileSize: bigint;
  /**
   * @generated from protobuf field: int64 thumbnailSize = 10
   */
  thumbnailSize: bigint;
}
// ...
export interface RecordMetaData {
  /**
   * @generated from protobuf field: bytes recordUid = 1
   */
  recordUid: Uint8Array;
  /**
   * @generated from protobuf field: bool owner = 2
   */
  owner: boolean;
  /**
   * @generated from protobuf field: bytes recordKey = 3
   */
  recordKey: Uint8Array;
  /**
   * @generated from protobuf field: Records.RecordKeyType recordKeyType = 4
   */
  recordKeyType: RecordKeyType;
  /**
   * @generated from protobuf field: bool canShare = 5
   */
  canShare: boolean;
  /**
   * @generated from protobuf field: bool canEdit = 6
   */
  canEdit: boolean;
  /**
   * @generated from protobuf field: bytes ownerAccountUid = 7
   */
  ownerAccountUid: Uint8Array;
  /**
   * @generated from protobuf field: int64 expiration = 8
   */
  expiration: bigint;
  /**
   * @generated from protobuf field: Records.TimerNotificationType expirationNotificationType = 9
   */
  expirationNotificationType: TimerNotificationType;
  /**
   * @generated from protobuf field: string ownerUsername = 10
   */
  ownerUsername: string;
}
// ...
export interface SharedFolderRecord {
  /**
   * @generated from protobuf field: bytes sharedFolderUid = 1
   */
  sharedFolderUid: Uint8Array;
  /**
   * @generated from protobuf field: bytes recordUid = 2
   */
  recordUid: Uint8Array;
  /**
   * @generated from protobuf field: bytes recordKey = 3
   */
  recordKey: Uint8Array;
  /**
   * @generated from protobuf field: bool canShare = 4
   */
  canShare: boolean;
  /**
   * @generated from protobuf field: bool canEdit = 5
   */
  canEdit: boolean;
  /**
   * @generated from protobuf field: bytes ownerAccountUid = 6
   */
  ownerAccountUid: Uint8Array;
  /**
   * @generated from protobuf field: int64 expiration = 7
   */
  expiration: bigint;
  /**
   * @generated from protobuf field: bool owner = 8
   */
  owner: boolean;
  /**
   * @generated from protobuf field: Records.TimerNotificationType expirationNotificationType = 9
   */
  expirationNotificationType: TimerNotificationType;
  /**
   * @generated from protobuf field: string ownerUsername = 10
   */
  ownerUsername: string;
  /**
   * @generated from protobuf field: bool rotateOnExpiration = 11
   */
  rotateOnExpiration: boolean;
}
// ...
export enum CacheStatus {
  /**
   * @generated from protobuf enum value: KEEP = 0;
   */
  KEEP = 0,
  /**
   * @generated from protobuf enum value: CLEAR = 1;
   */
  CLEAR = 1,
}
// ...
export const SyncDownResponse = new SyncDownResponse$Type();
// @generated message type with reflection information, may provide speed optimized methods
class UserFolder$Type extends MessageType<UserFolder> {
  constructor() {
    super("Vault.UserFolder", [
      { no: 1, name: "folderUid", kind: "scalar", T: 12 /*ScalarType.BYTES*/ },
      { no: 2, name: "parentUid", kind: "scalar", T: 12 /*ScalarType.BYTES*/ },
      { no: 3, name: "userFolderKey", kind: "scalar", T: 12 /*ScalarType.BYTES*/ },
      { no: 4, name: "keyType", kind: "enum", T: () => ["Records.RecordKeyType", RecordKeyType] },
      {
        no: 5,
        name: "revision",
        kind: "scalar",
        T: 3 /*ScalarType.INT64*/,
        L: 0 /*LongType.BIGINT*/,
      },
      { no: 6, name: "data", kind: "scalar", T: 12 /*ScalarType.BYTES*/ },
    ]);
  }
  create(value?: PartialMessage<UserFolder>): UserFolder {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.folderUid = new Uint8Array(0);
    message.parentUid = new Uint8Array(0);
    message.userFolderKey = new Uint8Array(0);
    message.keyType = 0;
    message.revision = BigInt(0);
    message.data = new Uint8Array(0);
    if (value !== undefined) reflectionMergePartial<UserFolder>(this, message, value);
    return message;
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: UserFolder,
  ): UserFolder {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* bytes folderUid */ 1:
          message.folderUid = reader.bytes();
          break;
        case /* bytes parentUid */ 2:
          message.parentUid = reader.bytes();
          break;
        case /* bytes userFolderKey */ 3:
          message.userFolderKey = reader.bytes();
          break;
        case /* Records.RecordKeyType keyType */ 4:
          message.keyType = reader.int32();
          break;
        case /* int64 revision */ 5:
          message.revision = reader.int64().toBigInt();
          break;
        case /* bytes data */ 6:
          message.data = reader.bytes();
          break;
        default:
          let u = options.readUnknownField;
          if (u === "throw")
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );
          let d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d,
            );
      }
    }
    return message;
  }
  internalBinaryWrite(
    message: UserFolder,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* bytes folderUid = 1; */
    if (message.folderUid.length) writer.tag(1, WireType.LengthDelimited).bytes(message.folderUid);
    /* bytes parentUid = 2; */
    if (message.parentUid.length) writer.tag(2, WireType.LengthDelimited).bytes(message.parentUid);
    /* bytes userFolderKey = 3; */
    if (message.userFolderKey.length)
      writer.tag(3, WireType.LengthDelimited).bytes(message.userFolderKey);
    /* Records.RecordKeyType keyType = 4; */
    if (message.keyType !== 0) writer.tag(4, WireType.Varint).int32(message.keyType);
    /* int64 revision = 5; */
    if (message.revision !== BigInt(0)) writer.tag(5, WireType.Varint).int64(message.revision);
    /* bytes data = 6; */
    if (message.data.length) writer.tag(6, WireType.LengthDelimited).bytes(message.data);
    let u = options.writeUnknownFields;
    if (u !== false) (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message Vault.UserFolder
 */
export const UserFolder = new UserFolder$Type();
// ...
export const SharedFolder = new SharedFolder$Type();
// ...
export const Record = new Record$Type();
// ...
export const RecordMetaData = new RecordMetaData$Type();
// ...
export const SharedFolderRecord = new SharedFolderRecord$Type();